+++ packages/webb-ui-components/.storybook/main.js
@@ -10,6 +10,8 @@ module.exports = {
     "@storybook/addon-links",
     "@storybook/addon-essentials",
     "@storybook/addon-interactions",
+    "@storybook-addon-react-router-v6",
+    "@storybook/theming",
     {
       name: '@storybook/addon-postcss',
         options: {
