+++ packages/webb-ui-components/.storybook/manager.js
@@ -1,0 +1,8 @@
+// .storybook/manager.js
+
+import { addons } from '@storybook/addons';
+import webbTheme from './webbTheme';
+
+addons.setConfig({
+  theme: webbTheme,
+});
