+++ packages/webb-ui-components/src/stories/molecules/AmountInput.stories.jsx
@@ -1,0 +1,25 @@
+import React from 'react';
+
+import { AmountInput } from '@webb-dapp/webb-ui-components/components';
+
+// More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
+export default {
+  title: 'Design System/Molecules/AmountInput',
+  component: AmountInput,
+  // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
+  // argTypes: {
+  //   type: { amount: 'string', required: false },
+  // },
+};
+
+// More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
+const Template = (args) => <AmountInput {...args}/>;
+
+export const Default = Template.bind({});
+// More on args: https://storybook.js.org/docs/react/writing-stories/args
+Default.args = {
+  title: 'md',
+  info: '',
+  variant: 'body4',
+  titleComponent: 'span'
+};
