+++ packages/webb-ui-components/src/stories/molecules/Avatar.stories.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
-import { Logo } from '@webb-dapp/webb-ui-components/components';
+import { Avatar } from '@webb-dapp/webb-ui-components/components';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Molecules/Logo',
-  component: Logo,
+  title: 'Design System/Molecules/Avatar',
+  component: Avatar,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
 //   argTypes: {
 //     backgroundColor: { control: 'color' },
@@ -13,26 +13,24 @@ export default {
 };
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template = (args) => <Logo {...args}/>;
+const Template = (args) => <Avatar {...args}/>;
 
 export const Default = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
 Default.args = {
+  key: '2',
   size: 'md',
+  src:  "./assets/colors.svg",
+  sourceVariant: 'address'
 };
 
-export const Large = Template.bind({});
-// More on args: https://storybook.js.org/docs/react/writing-stories/args
-Large.args = {
-  size: 'lg',
-};
+// export const Large = Template.bind({});
+// // More on args: https://storybook.js.org/docs/react/writing-stories/args
+// Large.args = {
+//   size: 'lg',
+// };
 
-export const Medium = Template.bind({});
-Medium.args = {
-  size: 'md',
-};
-
-export const Small = Template.bind({});
-Small.args = {
-    size: 'sm',
-  };
+// export const Medium = Template.bind({});
+// Medium.args = {
+//   size: 'md',
+// };
