+++ packages/webb-ui-components/src/stories/molecules/Divider.stories.tsx
@@ -4,7 +4,7 @@ import { Divider } from '@webb-dapp/webb-ui-components/components';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Molecules/Divider',
+  title: 'Design System/Molecules/Divider',
   component: Divider,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
 //   argTypes: {
