+++ packages/webb-ui-components/src/stories/molecules/TimeProgress.stories.jsx
@@ -1,11 +1,11 @@
-import React, { useState } from 'react';
-
-import { FixedAmount } from '@webb-dapp/webb-ui-components/components';
+import React from 'react';
+import { randRecentDate, randSoonDate } from '@ngneat/falso';
+import { TimeProgress } from '@webb-dapp/webb-ui-components/components';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Molecules/FixedAmount',
-  component: FixedAmount,
+  title: 'Design System/Molecules/TimeProgress',
+  component: TimeProgress,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
 //   argTypes: {
 //     backgroundColor: { control: 'color' },
@@ -13,15 +13,11 @@ export default {
 };
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template = (args) => <FixedAmount {...args} />;
+const Template = (args) => <TimeProgress {...args}/>;
 
-export const Default = Template.bind({});
-
-const handleChange = (nextVal) => setValue(nextVal);
+export const Primary = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
-Default.args = {
- info: 'Fix amount',
- values: [0.1, 0.25, 0.5, 1],
- value: '0.1',
- onChange: {handleChange}
+Primary.args = {
+    startTime: randRecentDate(),
+    endTime: randSoonDate()
 };
