+++ packages/webb-ui-components/src/stories/molecules/TokenInput.stories.jsx
@@ -4,7 +4,7 @@ import { TokenInput } from '@webb-dapp/webb-ui-components/components';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Molecules/TokenInput',
+  title: 'Design System/Molecules/TokenInput',
   component: TokenInput,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
 //   argTypes: {
